Use LongStream instead of IntStream and casts

+++ src/main/java/com/rehuapro/aoc2023/day6/Day6_1.java
@@ -7,12 +7,13 @@ import java.nio.file.Path;
 import java.util.*;
 import java.util.function.BinaryOperator;
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 public class Day6_1 {
 
     record Race(long time, long distance) {
         public long waysToWin() {
-            return IntStream.rangeClosed(0, (int) time)
+            return LongStream.rangeClosed(0, time)
                     .filter(i -> i * (time - i) > distance)
                     .count();
         }
